pyproject.toml
[tool.poetry]
name = "hw-27"
version = "0.1.0"
description = ""
authors = ["wooowno <sorrryks@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_27"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
psycopg2 = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
